CRUD Funcionando
Sem muito tempo, pra verificar se havia mais algum bug

+++ Teste/Utils/JsonBD.cs
@@ -1,14 +1,15 @@
 ﻿using Newtonsoft.Json;
-using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using Teste.Models;
 
 namespace Teste.Utils
 {
-    public static class JsonBD
+    public static class JsonBD<T> where T : ModelBase
     {
         // Funcionando
-        public static List<T> ObterArquivo<T>(string caminhoDoArquivo)
+        public static List<T> ObterArquivo(string caminhoDoArquivo)
         {
             List<T> lista = new List<T>();
 
@@ -24,14 +25,43 @@ namespace Teste.Utils
             return lista;
         }
 
+        public static T Obter(string caminhoDoArquivo, int id)
+        {
+            List<T> lista = new List<T>();
+
+            if (File.Exists(caminhoDoArquivo))
+            {
+                string json = File.ReadAllText(caminhoDoArquivo);
+                lista = JsonConvert.DeserializeObject<List<T>>(json);
+                return lista[id];
+            }
+            return lista.FirstOrDefault();
+        }
+
         // Funcionando: Inserir um objeto
-        public static void Inserir<T>(string caminhoDoArquivo, T domain)
+        public static void Inserir(string caminhoDoArquivo, T domain)
         {
             if (File.Exists(caminhoDoArquivo))
             {
                 List<T> lista = new List<T>();
                 string json = File.ReadAllText(caminhoDoArquivo);
                 lista = JsonConvert.DeserializeObject<List<T>>(json);
+                if (lista.Count > 0)
+                {
+                    int num = lista.ElementAt(0).Id;
+                    for (int i = 0; i < lista.Count; i++)
+                    {
+                        if (num < lista[i].Id)
+                        {
+                            num = lista[i].Id;
+                        }
+                    }
+                    domain.Id = (num + 1);
+                }
+                else
+                {
+                    domain.Id = 0;
+                }
                 lista.Add(domain);
                 json = JsonConvert.SerializeObject(lista, Formatting.Indented);
                 File.WriteAllText(caminhoDoArquivo, json);
@@ -39,7 +69,7 @@ namespace Teste.Utils
         }
 
         // Funcionando: Inserir um objeto
-        public static void InserirLista<T>(string caminhoDoArquivo, List<T> domain)
+        public static void InserirLista(string caminhoDoArquivo, List<T> domain)
         {
             if (File.Exists(caminhoDoArquivo))
             {
@@ -52,7 +82,7 @@ namespace Teste.Utils
             }
         }
 
-        public static void Alterar<T>(string caminhoDoArquivo, int id, T domain)
+        public static void Alterar(string caminhoDoArquivo, int id, T domain)
         {
             if (File.Exists(caminhoDoArquivo))
             {
@@ -66,7 +96,7 @@ namespace Teste.Utils
             }
         }
 
-        public static void Excluir<T>(string caminhoDoArquivo, int id)
+        public static void ExcluirPorId(string caminhoDoArquivo, int id)
         {
             if (File.Exists(caminhoDoArquivo))
             {
@@ -74,6 +104,20 @@ namespace Teste.Utils
                 string json = File.ReadAllText(caminhoDoArquivo);
                 List<T> lista = JsonConvert.DeserializeObject<List<T>>(json);
                 lista.RemoveAt(id);
+                // Salva o arquivo de novo
+                json = JsonConvert.SerializeObject(lista, Formatting.Indented);
+                File.WriteAllText(caminhoDoArquivo, json);
+            }
+        }
+
+        public static void Excluir(string caminhoDoArquivo, T domain)
+        {
+            if (File.Exists(caminhoDoArquivo))
+            {
+                // Transforma o arquivo em json e remove
+                string json = File.ReadAllText(caminhoDoArquivo);
+                List<T> lista = JsonConvert.DeserializeObject<List<T>>(json);
+                lista.Remove(domain);
                 // Salva o arquivo de novo
                 json = JsonConvert.SerializeObject(lista, Formatting.Indented);
                 File.WriteAllText(caminhoDoArquivo, json);
